devtools/client/webide/test/test_autoconnect_runtime.html

<!DOCTYPE html>

<html>

  <head>
    <meta charset="utf8">
    <title></title>

    <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript" src="chrome://mochikit/content/chrome-harness.js"></script>
    <script type="application/javascript" src="head.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  </head>

  <body>

    <script type="application/javascript">
      window.onload = function() {
        SimpleTest.waitForExplicitFinish();

        (async function() {
          DebuggerServer.init();
          DebuggerServer.registerAllActors();

          let win = await openWebIDE();
          let docRuntime = getRuntimeDocument(win);

          let fakeRuntime = {
            type: "USB",
            connect: function(connection) {
              is(connection, win.AppManager.connection, "connection is valid");
              connection.host = null; // force connectPipe
              connection.connect();
              return Promise.resolve();
            },

            get id() {
              return "fakeRuntime";
            },

            get name() {
              return "fakeRuntime";
            }
          };
          win.AppManager.runtimeList.usb.push(fakeRuntime);
          win.AppManager.update("runtime-list");

          let panelNode = docRuntime.querySelector("#runtime-panel");
          let items = panelNode.querySelectorAll(".runtime-panel-item-usb");
          is(items.length, 1, "Found one runtime button");

          let connectionsChanged = waitForConnectionChange("opened", 2);
          items[0].click();

          ok(win.document.querySelector("window").className, "busy", "UI is busy");
          await win.UI._busyPromise;

          await connectionsChanged;
          is(Object.keys(DebuggerServer._connections).length, 2, "Connected");

          connectionsChanged = waitForConnectionChange("closed", 2);

          await nextTick();
          await closeWebIDE(win);

          await connectionsChanged;
          is(Object.keys(DebuggerServer._connections).length, 0, "Disconnected");

          connectionsChanged = waitForConnectionChange("opened", 2);

          win = await openWebIDE();

          win.AppManager.runtimeList.usb.push(fakeRuntime);
          win.AppManager.update("runtime-list");

          await waitForUpdate(win, "runtime-targets");

          await connectionsChanged;
          is(Object.keys(DebuggerServer._connections).length, 2, "Automatically reconnected");

          await win.Cmds.disconnectRuntime();

          await closeWebIDE(win);

          DebuggerServer.destroy();

          SimpleTest.finish();
        })();
      };
    </script>
  </body>
</html>
